penyesuaian form data mutu

+++ database/migrations/2025_07_14_212714_add_column_tingkat_kepuasan_to_kuisoner_kepuasan_table.php
@@ -13,9 +13,11 @@ class AddColumnTingkatKepuasanToKuisonerKepuasanTable extends Migration
      */
     public function up()
     {
-        Schema::table('kuisoner_kepuasan', function (Blueprint $table) {
-            $table->string('tingkat_kepuasan')->nullable()->after('saran');
-        });
+        if (!Schema::hasColumn('kuisoner_kepuasan', 'tingkat_kepuasan')) {
+            Schema::table('kuisoner_kepuasan', function (Blueprint $table) {
+                $table->string('tingkat_kepuasan')->nullable()->after('saran');
+            });
+        }
     }
 
     /**
@@ -25,8 +27,10 @@ class AddColumnTingkatKepuasanToKuisonerKepuasanTable extends Migration
      */
     public function down()
     {
-        Schema::table('kuisoner_kepuasan', function (Blueprint $table) {
-            $table->dropColumn('tingkat_kepuasan');
-        });
+        if (Schema::hasColumn('kuisoner_kepuasan', 'tingkat_kepuasan')) {
+            Schema::table('kuisoner_kepuasan', function (Blueprint $table) {
+                $table->dropColumn('tingkat_kepuasan');
+            });
+        }
     }
 }
